Revert "Fixed compilation under MinGW."
This reverts commit 926f3c969123683656dc922aa9bac563149a4687.

+++ libraries/utilities/include/mcrl2/utilities/detail/unordered_set_implementation.h
@@ -203,7 +203,7 @@ void MCRL2_UNORDERED_SET_CLASS::rehash(std::size_t number_of_buckets)
   // Update the number of mutexes.
   if constexpr (!EnableLockfreeInsertion)
   {
-    m_bucket_mutexes = std::vector<std::mutex>(std::max(number_of_buckets / BucketsPerMutex, static_cast<std::size_t>(1ul)));
+    m_bucket_mutexes = std::vector<std::mutex>(std::max(number_of_buckets / BucketsPerMutex, 1ul));
   }
 
   // Create one bucket list for all elements in the hashtable.
